Update Take-a-break.py
Here we used Tkinter library, to make the source code even more interactive with the user. The Tkinter library is used to add the GUI to the program and improve the interaction with the program. We can now display a message box with a message as "Time to relax!" once the timer comes to end.